requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
twilio==8.13.0
google-generativeai==0.3.2
pypdf2==3.0.1
pydantic==2.5.3
httpx==0.26.0
python-multipart==0.0.6
websockets==12.0
aiofiles==23.2.1
jinja2==3.1.3
